refactor offline init and cih init
