+++ crates/osutils/src/osrelease.rs
@@ -5,7 +5,7 @@ use const_format::formatcp;
 use log::trace;
 use serde::{Deserialize, Deserializer};
 
-use crate::{path, special_release::SpecialRelease};
+use crate::path;
 
 /// Absolute path to the /etc/os-release file.
 pub const OS_RELEASE_PATH: &str = "/etc/os-release";
@@ -66,20 +66,20 @@ impl OsRelease {
                     })
                     .unwrap_or_default(),
             ),
-            Some("special_legacy") => Distro::Special(SpecialRelease::SpecialLegacy),
-            Some("special") => Distro::Special(
-                self.version_id
-                    .as_deref()
-                    .map(|v| {
-                        if v.starts_with("1.") {
-                            SpecialRelease::Special1
-                        } else {
-                            trace!("Unknown Special release: {v}");
-                            SpecialRelease::Other
-                        }
-                    })
-                    .unwrap_or_default(),
-            ),
+            // Some("special_legacy") => Distro::Special(SpecialRelease::SpecialLegacy),
+            // Some("special") => Distro::Special(
+            //     self.version_id
+            //         .as_deref()
+            //         .map(|v| {
+            //             if v.starts_with("1.") {
+            //                 SpecialRelease::Special1
+            //             } else {
+            //                 trace!("Unknown Special release: {v}");
+            //                 SpecialRelease::Other
+            //             }
+            //         })
+            //         .unwrap_or_default(),
+            // ),
             _ => Distro::Other,
         }
     }
@@ -193,7 +193,7 @@ impl ExtensionRelease {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum Distro {
     AzureLinux(AzureLinuxRelease),
-    Special(SpecialRelease),
+    // Special(SpecialRelease),
     Other,
 }
