+++ crates/trident/src/init/mod.rs
@@ -1,0 +1,2 @@
+pub mod cih;
+pub mod offline;
