+++ crates/trident/src/init/offline/lazy_lsblk.json
