+++ crates/trident/src/init/offline/lsblk.json
